feat!: Add macOS support
This commit makes the plugin work on macOS with shared code as described in the
["Shared iOS and macOS implementations"](https://docs.flutter.dev/packages-and-plugins/developing-packages#shared-ios-and-macos-implementations)
section of Flutter plugin docs.

The following changes were required to achieve this:
* Delete the `ios` directory and move everything to `darwin`
* Increase the lowest supported Flutter version to 3.7.0, since older
  versions don't support this shared structure
* Add conditional imports for `Flutter` or `FlutterMacOS` based on
  platform
* Delete the iOS-native `fluttertoast` package from the example app in
  favor of a built-in Flutter snackbar

Tested on iPhone 11 with iOS 16.6 & M1 MacBook Air with macOS 13.2.1

This is a breaking change for folks who use older Flutter versions,
otherwise nothing's changed.

Closes #6